Don't rewind the build if invocation id stays the same (#18670)
Invocation id (`BUILD_ID` in code) is expected to be different for each command. When retrying the build, we rely on Bazel generating a new invocation id for a new attempt. However, if flag `--invocation_id` is set, Bazel just uses the provided value instead of generating a new one. In this case, invocation id stays the same among multiple attempts which could cause issues like https://github.com/bazelbuild/bazel/issues/18547.

This PR fixes that by not retrying the build if the invocation id is same to previous attempt. Also updated the doc to point this requirement out.

Closes #18591.

PiperOrigin-RevId: 539946840
Change-Id: I6ae85ea923b0fdbff97fe2e44e36995f0205f8a1