Prepare tests for Starlark JavaInfo
In order to correctly match stringified JavaInfo in `JavaRoundTripTest`, we need to strip certain fields anywhere in the content. This is beacuse instances of Starlark providers also expose the `to_json` and `to_proto` methods from StructApi. Native provider instances (like JavaInfo) generally do not because they use a closer ancestor class (such as `JavaInfoApi`) to define their public API. Additionally we strip any fields beginning with "_" which are private and also not available in the native provider's API.

PiperOrigin-RevId: 540375596
Change-Id: Ie0170507a34569e584bebe92530c700d24366edd