Make it possible to use BwoB with Skymeld.
Particularly the downloading behavior during a build. We achieve this by incrementally building the list of artifacts to download, to make up for the lack of a complete `AnalysisResult` in skymeld.

This however brings about a regression in incremental builds: instead of dirtying just the artifacts of top-level targets, we would need to dirty all action nodes in the Skyframe graph. We'd for sure be doing more work, but likely the other layers of caching would help alleviate this. This should be nonetheless addressed in a follow up CL.

PiperOrigin-RevId: 540547330
Change-Id: I8a2639fee135afb629b50d28846e8a3439612f56