Make `JavaGenJarsProvider` an interface implemented by `NativeJavaGenJarsProvider`
Purely mechanical change with no other changes.

This is needed for Starlark JavaInfo because this provider holds references to other JavaGenJarsProvider instances which makes eager, recursive copying into native instances too expensive. The plan is to add a StarlarkJavaGenJars class (in a following change) that only wraps the StarlarkInfo and extracts data lazily. This is similar to what was done for `ProtoInfo`.

Eventually, once there is no construction of JavaInfo natively, we will be left with only the wrapper class.

PiperOrigin-RevId: 540560827
Change-Id: I6f204c6572cc4c913639002e6887d261bd4b8363