Make `test_interrupted_children_waited` faster.
It used to rely on `sleep()` for synchronization. That was slow *and* theoretically incorrect.

There is still a `sleep()` invocation, but it only matters if the test fails. Its role is to not let the test timeout, but fail faster.

I also attempted to verify that the repo process finishes by adding a `trap ... EXIT` to `waiter.sh`, but that does not work for some reason. I did not investigate further.

RELNOTES: None.
PiperOrigin-RevId: 540575714
Change-Id: I48781e545f41d99a38887e0fd27ddad5937fb235