ApiVersion: switch to use ClientApiVersion
ApiVersion is a class that wrap around RemoteApi's SemVer proto.
We have been relying on ApiVersion.current to determine the current
version of RemoteApi that Bazel supports.

However, to start supporting newer versions of RemoteApi while providing
backward compatibility with different server implementations, we don't
want to reduce our support to a single version. Instead, we need a range
of versions that Bazel (client-side) supports to match with the range of
versions the server supports, which is provided via RemoteApi's
ServerCapabilities.

Create ClientApiVersion to supply this version range logic and move
support status check from ApiVersion to the new class.

Enhance ServerSupportedStatus with the inclusion of the highest
supported version, which could either be "null", in case of Unsupported
status, or is the highest supported version in case of Supported or
Deprecated status.

This work should enable #18270 in a future change. Previously, #18270
was reverted because backward incompatible with older RemoteApi
versions.

Closes #18647.

PiperOrigin-RevId: 540591460
Change-Id: I65fc07ec2d36a11ed512a980bcdf7356451e566a