Store reverse deps of size 2-4 as a `SkyKey[]` instead of an `ArrayList`.
Most nodes have a small number of reverse deps. There is already an optimization to store a single reverse dep bare. This change extends the optimization to forego an `ArrayList` wrapper (24 bytes) when there are 2-4 elements.

Additionally, this saves space that was being wasted on over-allocating the size of the backing array by calling `Lists.newArrayListWithExpectedSize(2)`, which per the Javadoc on that method, adds "an unspecified amount of padding" (also, the Javadoc discourages using the method altogether). The actual calculation is `5L + arraySize + (arraySize / 10)`, which means for an input of 2 we were allocating an array of size 7.

PiperOrigin-RevId: 540717655
Change-Id: I8d485f8ca463a4178e2844491928748d5f6c182c