Remove superfluous code in RemoteExecutionService.
The directories that are to contain spawn outputs have already been created before the spawn executes, so it's unnecessary for RemoteExecutionService to create them (however, tests must be amended to establish that precondition).

The comment incorrectly suggests this is a fix for #6260. The actual fix was 4392ba4.

In addition, fix a bug whereby materializing an output directory as a symlink would fail because the directory already exists.

PiperOrigin-RevId: 541623490
Change-Id: I25abd3b7b5fc068a84f41c86b0050c110443218f