Intern `EnvironmentCollection`.
`hashCode` and `equals` consider the order of keys in the map, since various methods are documented to return collections that match the order in which elements were added to the builder. The reference type of the backing map is changed to `ImmutableListMultimap` to make it more apparent that its equality already considers the order of values for a particular key.

Additionally, rewrite `getGroupedEnvironments` to pre-size the builder and use `forEach` to reduce allocations.

PiperOrigin-RevId: 541940702
Change-Id: I1661bcd2b626d420a4cd58a24c760f6bf129fb38