Support features in apple_common.link_multi_arch_binary
This will be used to convert some linkopts in apple rules to features, which in
turn will allow the crosstool to select linking options based on the kind of
linking we're doing (executable, bundle, relocatible, or dylib).

Note we couldn't pass in a feature_configuration from the starlark, because it
is before the split transition.  We need to pass feature strings into native
code, and have it apply those features after the transition.

PiperOrigin-RevId: 542032473
Change-Id: I0b769744db5d3286e7af8d9af173151d0b838738