Update remote worker to modify mtime of referenced blobs during FindMissingBlobs call
to simulate lease extension.

The purpose of this CL is to setup a testbed for incoming changes that implement lease extension in Bazel. It doesn't try to fully simulate the lease extension because it currently only touches files referenced by FindMissingBlobs. In practice, there are more APIs to cover, e.g. GetActionResult.

PiperOrigin-RevId: 542197583
Change-Id: I6467ef1ba9c9f1fb830dcbfa1cd4fbb63567fc32