Fix order of linker inputs in cc_[shared_library|binary]
In https://github.com/bazelbuild/bazel/commit/c73900bb999bc8dae938fdff341aac3816bcf0ed the order of the linker inputs in the linking command line was fixed for common archives between different direct deps. Before that fix the direct deps were merged sequentially without taking into account dependencies (no topological sort). After the change we were doing a topological sort but it was wrong in a way.

The cc_shared_library graph is a parallel graph to the cc_library graph. The former decides partitions of the latter. From the top level cc_library every other cc_library is reachable even if it was already linked to a cc_shared_library in the dynamic dependencies. The depset that we were using for linker inputs only took into account the cc_libraries for doing a topological sort, it didn't take into account any pruning we might have done from the graph due to the presence of a shared library. After the depset was sorted we then filtered the linker inputs of cc_libraries that were already present in cc_shared_libraries in dynamic_deps.

This was wrong. A cc_shared_library should be treated as a black box and its dependencies should not have any influence on the final topological order of the linker inputs. This CL addresses that by taking into account pruned inputs.

In a first attempt this change re-implemented topological sorting using Kahn's algorithm but this was different to what is implemented by depset(order="topological") since in Kahn's the nodes without incoming edges are always processed first. This had the undesired effect that even when it was possible to respect the order of the deps as listed on the BUILD file and still produce a valid topological sort, Kahn's ignored the order. A second attempt (the one in this CL) re-builds a topological depset by changing the pruning algorithm to traverse the graph depth-first, then lets Bazel handle the sorting by flattening the depset.

This CL also fixes the order for linker_inputs in cc_binary which had a wrong order by listing shared libraries first followed by static linker inputs without any consideration for all of the above.

RELNOTES:none
PiperOrigin-RevId: 542208041
Change-Id: Icaf3b488f3208259ff0bb0eef3318bef582533c5