RemoteCache: keep ServerCapabilities instead of CacheCapabilities
In Remote Execution Api, ServerCapabilities includes CacheCapabilities
and ExecutionCapabilities and the Api Version metadata for version
compatibility check.

To helps check Bazel compatibility with Remote Cache implementations
that are using Remote Execution API, we would need access to the Api
Version metadata in ServerCapabilities.
So instead of storing just a subset of ServerCapabilities, expand
RemoteCache to store the whole ServerCapabilities instead.

This is a part of #18270.