Express starlark_doc_extract's dependencies in the target graph
We want to ensure that the source Starlark module label really is a label of an
input file (not e.g. a rule output, which we cannot load), and that we resolve any
aliases before using the label.

We also, unfortunately, need to add a deps attribute so that bazel query can list
starlark_doc_extract target's transitive dependencies. We intend to deps to be a
list of bzl_library targets - but cannot enforce this at the moment (and are forced
to read deps via DefaultInfo instead of StarlarkLibraryInfo) because bzl_library
and StarlarkLibraryInfo live in bazel_skylib, so a native rule cannot depend on
them.

We verify at analysis time that the (transitive) files-to-build of the (presumed-
to-be) bzl_library targets in deps are (1) input files (not rule outputs, which we
cannot load) and (2) a superset of the transitive loads of the source Starlark
module.

For collecting those transitive loads, we reuse the existing LoadGraphVisitor class
and visitLoadGraphRecursively() from Package - which means moving them from Package
to BazelModuleContext.

PiperOrigin-RevId: 542324998
Change-Id: I0898ce2b24d82d3cd13c4ea690c19065d29af8a7