Simplify the synchronization logic in RunfilesTreeUpdater.
Since the contents of a given runfiles tree and the decision of whether to materialize it are constant for a given build command, there's no need to update it more than once; further attempts should just wait for the first attempt to complete.

Also use a ConcurrentHashMap, which has better performance under contention than the current implementation.

PiperOrigin-RevId: 542369121
Change-Id: Ibdf1bd3173cb648f78234865715b7f87ccfeb727