Use direct delegation only instead of ProxyConfiguredTargetKey.
This is possible because rule transitions that actually change the configuration
are rare enough that using delegation in those instances adds very few nodes.

The other common cause of configuration change is with non-configurable
targets. The majority of these turn out to be in the same package as the
parent. This means that inspecting those targets locally doesn't break
the desired encapsulation so we can determine the null configuration
in advance.

PiperOrigin-RevId: 542442858
Change-Id: I5867286db14719a937d94c4e05d12b94fc561204