Make --action_cache_store_output_metadata a no-op.
Storing metadata for remote outputs in action cache allows Bazel to use action cache across server restarts for remote actions and the flag was used to determine whether Bazel should do that.

Since we always want to use it for BwoB and we never want to use it for other cases, this CL changes Bazel to make the decision based on output mode and make this flag an no-op.

PiperOrigin-RevId: 542522993
Change-Id: I5dfe355a0a5be213cad9921ea0ca3b3b9d2df91c