[Skymeld] Fix a race condition with the IncrementalArtifactConflictFinder when there's an analysis error.
When there's an analysis error, the incremental artifact conflict finder is shut
down [1]. This could create a race condition with the other Skyframe threads
which are evaluating the other BuildDriverKeys and requesting for conflict checking.

This resulted in the `RejectedExecutionException` that sometimes appeared with
an analysis error.

This CL adds some synchronizations to make sure that no new tasks is submitted
after the executor has been shut down.

PiperOrigin-RevId: 542782560
Change-Id: I5d6a9a1d98a1256910a9a1d92b67de9a36fb0120