Done operations must be reexecuted
Any operation with done == true as reported by the server is not
expected to change its result on subsequent waitExecution calls. To
properly retry, this action must be reexecuted, if it was truly
transient, to achieve a definitive result. Submit a transient status for
retry, disallow special behaviors for NOT_FOUND as covered by done
observation, and consider method type when handling operation streams.