Report remote execution messages as events
Currently the message is appended to a spawn's stderr, which pollutes
that output. It also means that whether a message is output is gated by
the `--output_filter` flag, which is surprising - metadata messages
don't feel like "output" in the same way `--output_filter` applies to.

Instead, emit the message as a top-level message to the terminal.