Delay connecting to the remote executor until executorInit
This change moves the majority of the code in the beforeCommand hook
of the RemoteModule to executorInit.  This makes the former validate
the configuration flags only, and postpones any attempts to connect
to the remote executor until it's actually needed.

The reason for delaying the connection to the executor is to ensure
that commands that do /not/ require it can run through completion
without errors nor pauses.

To make the diff of this change small, I've tried to leave the code
where it was by wrapping it in a new function that gets called from
executorInit.  The indirection is not necessary for anything else
though.

Fixes #18760.