Avoid discarding SRE state for IO cause
Unwrapping all StatusRuntimeExceptions in in ReferenceCountedChannel
when caused by IOException will discard critical tracing and
retriability. The Retrier evaluations may not see an SRE in the causal
chain, and presume it is invariably an unretriable exception. In
general, IOExceptions as SRE wrappers are unsuitable containers and are
routinely misued either for identification (grpc aware status), or
capture (handleInitError).