Reduce memory spent on progress messages of Java actions.
Uses the `%{output}` placeholder to lazily create the message with the JAR path instead of using the deprecated `OnDemandString` method.

There is a slight difference in the message: it will now contain the `blaze-out/k8-opt/` prefix, but this is more in line with progress messages from Starlark actions. Maintaining the exact same message (no prefix) is possible, but requires opening up the `SpawnAction` API a bit more, so hopefully the prefix is acceptable.

PiperOrigin-RevId: 543493741
Change-Id: Ib54b7cb09bdca3b1ac44949e0e10f14b3e582be8