Create the initial version of the packages_used rule
`packages_used` writes transitive SBOM data to a file for future
processing.  By itself, the rule is not useful. It produces input for
an SBOM generator (TBD in a future PR). That generator will take the
output from packages_used, and merge with the maven lock file to produce
a richer SBOM.