Create the initial version of the packages_used rule
`packages_used` writes transitive SBOM data to a file for future processing.  By itself, the rule is not useful. It produces input for an SBOM generator (TBD in a future PR). That generator will take the output from packages_used, and merge with the maven lock file to produce a richer SBOM.

Most of this will be moved to rules_license after we finish this proof-of-concept within the Bazel sources.

```
bazel build //tools/compliance:all
more bazel-bin/tools/compliance/bazel_packages.json
```

Closes #18782.

PiperOrigin-RevId: 543751564
Change-Id: I0cf0fcefa3c5477e6c3bdfe064693e13413887a0