Add SBOM generator - first cut
This adds an `sbom` rule and uses it to produce one for Bazel.

```
bazel build //tools/compliance:all
more bazel-bin/tools/compliance/bazel_sbom.json
```

This is no where near complete. It just establishes the rule and the helper to have an end to end path. One or more followup PRs will add more fields to the SBOM.

Closes #18789.

PiperOrigin-RevId: 543956376
Change-Id: I9ab593561e58cf4e6a4d2ba8ade3b00bcc947fe5