Fix remote execution tests
by replacing `mock` with `spy` because we rely on the real method and these methods are final.

Before mockito 5, it uses sublcass mockmaker which cannot mock final method/class. So when we call final methods on these mocked classes, we are calling into the real methods.

Now, it uses inline mockmaker which **can** mock final method/class. Calling final methods on these mocked classes will return `null` if we didn't provide the stub.

This change fixes that by using `spy`.

Closes #18800.

PiperOrigin-RevId: 544038459
Change-Id: Id75a5cb3b9d14e38d6bec918449e5aee671471eb