Fix SkyframeTests
Similar to https://github.com/bazelbuild/bazel/commit/56370834dc7dd0477d0436ae0a11fd9ab3271044, use `spy` instead of `mock` for the test that relies on real method.

The tests rely on `FileArtifactValue#addTo` which is a final method. Before Mockito 5, when we call this method on a mocked object, we call into the real method. With Mockito 5, we can mock the final method and now they are doing nothing which breaks the tests.

PiperOrigin-RevId: 544056698
Change-Id: Ie4340ee9cf2615928cc3df4b11be891ef9ddb3d2