Add INFO log lines to all call paths to `#handleAnalysisInvalidatingChange`.
This way we have post-facto debugging information for why we dropped the analysis cache. In particular, this is useful for when the cause is a different build configuration; Blaze already prints a helpful configuration diff to stderr but the user might not have saved that or noticed it (and even comparing the full set of options we already log to INFO is a bit tedious).

PiperOrigin-RevId: 544102574
Change-Id: Ia97ea099dea20ec97e38259db609867192fc65ff