Start to incorporate data from the maven lock into our SBOM
This works so far, but oh what hackery. Bazel uses a non-standard
way of referencing maven artifacts, so the external references we
see don't map back to the lock file directly. And the lock file is
a tricky beast to understand.

I am happy I have a few elements in place.  I will come back for
more in the next pass.