Support R.txt files generated with aapt2 link --package-id
ResourceProcessorBusyBox is being used in the Android platform build, which includes CTS tests that set `--package-id 0x80` when running aapt2 link.  This produces R.txt files with ids that are out of range of java integers, for example `int color blue 0x80020000`.  In an R.java file, javac interprets these as negative integers, but IntFieldInitializer throws a NumberFormatException.  Make IntFieldInitializer act like javac by parsing the number as a Long and then narrowing it to an int.

PiperOrigin-RevId: 544184014
Change-Id: Ida26d00a4c487830a8b35bbdf58dd9a5c5411adc