Report dev/non-dev deps imported via non-dev/dev usages
If a repository generated by a module extension corresponds to a dev tag
but is imported on non-dev usage (or vice versa), the build can fail if
the module is used by other modules. `ModuleExtensionMetadata` already
generated fixup commands for this case, but now also explains the
situation and impact.