Expose metadata_files parameter in coverage_common.instrumented_files_info
Allows all rules to provide additional metadata files to the
InstrumentedFilesInfo provider.

It was always a little odd that the metadata_files attribute on the provider
was exposed, but there was no way of populating it from Starlark.

The other parameters are still considered too tied up with internal details at
the moment to be exposed at this time.

RELNOTES: coverage_common.instrumented_files_info now has a metadata_files argument
PiperOrigin-RevId: 545391150
Change-Id: I3febdaf23c82d60c32b241a3935da27fd086fb7a