Delete top-level rule transition special case code.
This is no longer needed now that configured targets perform their own
rule transitions.

* Decomposes configuration logging differently. The
BuildConfigurationValue is often not available where the transitions
occur, sometimes never being computed at all, as is the case of an
attribute transition followed immediately with a rule transition. In
the case of a child-side rule transition, the BuildConfigurationValue
isn't computed locally. It's computed instead by the delegate,
which doesn't know the parent checksum. To work around this,
logs ConfigurationTransitionEvents as a pair of parent and child
hashes, then logs ConfigurationValueEvents as they are created in
BuildConfigurationFunction. This data can be joined in InvocationLogger
to recover what used to be in the ConfigRequestedEvents.

* There were a couple of places where additional filtering was needed,
similar to filters added for aquery/cquery. SequencedSkyframeExecutor
has a couple dump methods where nodes that only do delegation need
to be filtered out. The same filter is also needed in
ActionLookupValuesTraversal.accumulate.

PiperOrigin-RevId: 545687484
Change-Id: I42060c6f399870ee0abd08dffa1364467f2a4098