Fix race condition for tests with `--experimental_cancel_concurrent_tests`.
When `--experimental_cancel_concurrent_tests` is set, Bazel will cancel concurrently running tests on the first successful run by interrupting threads these tests running on.

Normally, if the test runner hasn't started running the test, or is waiting for the test to finish, the interruption can cancel that blocking operation and clear the interrupted status of the thread.

However, current code doesn't handle the case when the tests to be cancelled are already finished which leads to two problems:
  1. The interrupted status is propagated to skyframe causing the entire invocation to be cancelled.
  2. The assumption that "no test log is generated if the test was cancelled" is no longer true.

This CL fixes that by:
  1. Inside `AttemptGroup#cancelOthers`, we only cancel others if this group hasn't been cancelled to prevent the winning thread got cancelled by other threads that are already finished when receiveing the interruption.
  2. Check thread's interrupted status after the test attempt is finished.
  3. Remove the assertion that "no test log is generated if the test was cancelled".

PiperOrigin-RevId: 546214028
Change-Id: I3fc8d4e446ccabdb4c7b1794a3aeb8052912a044