Support passing LLVM MemProf (memory profile) to CppCompile
Adds support for supplying a MemProf profile, which can be supplied to LLVM to
automatically hint allocations based on profiled memory behavior. The patch
adds Blaze flag --memprof_profile=<label>, along with a memprof_profile rule
that allows specifying a profile either by absolute path or label.

The profile can either be an LLVM memprof profile with a .profdata extension,
or a zipfile containing one. The profile is symlinked and optionally unzipped,
and set in the memprof_profile_path build variable, which can be used to pass
the profile to the compiler via a feature.

PiperOrigin-RevId: 546311130
Change-Id: I896898bf3bddcf42214cde3f3f27d2ed18de6b26