Fix null event issue
If a change "only" occurred in a module extension, then the value of ModuleResolutionEvent is null -the skyvalue is cached and the event was never sent- then "combineModuleExtensions" function would crash with null pointer exception while trying to get the old usages.

In this case we can just re-add the old module extensions from the lockfile because if the module resolution is the same, this means the usage didn't change.

PiperOrigin-RevId: 546821911
Change-Id: Ie685cbab654d1c41403aebd31ddad91033be1d56