Enable C++ branch coverage when using gcov 8 or later.
Branch coverage was disabled due to a gcov bug, but this was fixed for
GCC 8: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=84879.

It doesn't make sense to disable all C++ branch coverage for
everyone because of a bug in an old GCC version.

RELNOTES: Enable C++ branch coverage if gcov version is 8 or newer.