Sort documentable entities (including rule attributes) in markdown output
starlark_doc_extract, unlike the legacy documentation extractor, produces a
proto with documentable entities in the same order as encountered in the
source. Thus, to be compatible with both the legacy and the new extractor,
the markdown renderer needs to sort rules/providers/functions/aspects, and
the fields in a given rule, before emitting output.

Since we are messing with the sort order, take the opportunity to sort rule
attributes in the same order as in the documentation for native rules in
Bazel docs. (Legacy Stardoc behavior was to sort everything except "name"
in simple alphabetical order. This means some golden tests will need to be
updated.)

PiperOrigin-RevId: 546966417
Change-Id: I69484168b1584aca662fd285652b3fbb3c1fef09