Enable C++ branch coverage when using gcov 8 or later.
Branch coverage was disabled due to a gcov bug, but this was fixed for
GCC 8: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=84879.

It doesn't make sense to disable all C++ branch coverage for everyone
because of a bug in an old GCC version.

Closes #18879.

RELNOTES: Enable C++ branch coverage if gcov version is 8 or newer.
PiperOrigin-RevId: 547165873
Change-Id: I581d9902dc73910f686155548c3bf94c7a1d9207