[6.3.0] Lockfile updates (#18894)
* Update lockfile function exception

Add new exception for this function to handle any syntax errors or missing data within the lockfile

Related: https://github.com/bazelbuild/bazel/issues/18455
PiperOrigin-RevId: 538144339
Change-Id: I82160f3bff6598c26b3f99c824fe85ea86086c1f

* Update lockfile writing logic to be event triggered.

The writing will only happen at the end of resolution in the after command function (at this point, the module and all needed module extensions are resolved). Which solves the problem of reading and writing into the lockfile multiple times in one invocation.

PiperOrigin-RevId: 540552139
Change-Id: I4a78412a388bde2ff7949d119831318c40d49047

# Conflicts:
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/BazelDepGraphFunction.java
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/BazelLockFileFunction.java
#	src/test/java/com/google/devtools/build/lib/bazel/bzlmod/BazelLockFileFunctionTest.java

* Add module extension to lockfile

PiperOrigin-RevId: 543445157
Change-Id: Ib32a2c9fdc22c5b228d78141fc9648b3ef5edf7d

# Conflicts:
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/GsonTypeAdapterUtil.java
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/SingleExtensionEvalFunction.java

* fixes

* Fix tests

* Enable lockfile by default and fix warning

PiperOrigin-RevId: 545927412
Change-Id: I34e8531b8e396ccdfe0eecbee22cb68f76f969fc

# Conflicts:
#	src/test/java/com/google/devtools/build/lib/analysis/RunfilesRepoMappingManifestTest.java
#	src/test/java/com/google/devtools/build/lib/bazel/bzlmod/BazelDepGraphFunctionTest.java
#	src/test/java/com/google/devtools/build/lib/bazel/bzlmod/BazelModuleResolutionFunctionTest.java
#	src/test/java/com/google/devtools/build/lib/query2/testutil/SkyframeQueryHelper.java
#	src/test/java/com/google/devtools/build/lib/rules/starlarkdocextract/StarlarkDocExtractTest.java

* Add lockfile documentation

PiperOrigin-RevId: 546085523
Change-Id: If287e6e143f8858d185f83a1630275520db65e44

# Conflicts:
#	site/en/_book.yaml

* Fix null event issue

If a change "only" occurred in a module extension, then the value of ModuleResolutionEvent is null -the skyvalue is cached and the event was never sent- then "combineModuleExtensions" function would crash with null pointer exception while trying to get the old usages.

In this case we can just re-add the old module extensions from the lockfile because if the module resolution is the same, this means the usage didn't change.

PiperOrigin-RevId: 546821911
Change-Id: Ie685cbab654d1c41403aebd31ddad91033be1d56

* Fix tests

* Update lockfile function exception

Add new exception for this function to handle any syntax errors or missing data within the lockfile

Related: https://github.com/bazelbuild/bazel/issues/18455
PiperOrigin-RevId: 538144339
Change-Id: I82160f3bff6598c26b3f99c824fe85ea86086c1f

* Update lockfile writing logic to be event triggered.

The writing will only happen at the end of resolution in the after command function (at this point, the module and all needed module extensions are resolved). Which solves the problem of reading and writing into the lockfile multiple times in one invocation.

PiperOrigin-RevId: 540552139
Change-Id: I4a78412a388bde2ff7949d119831318c40d49047

# Conflicts:
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/BazelDepGraphFunction.java
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/BazelLockFileFunction.java
#	src/test/java/com/google/devtools/build/lib/bazel/bzlmod/BazelLockFileFunctionTest.java

* Add module extension to lockfile

PiperOrigin-RevId: 543445157
Change-Id: Ib32a2c9fdc22c5b228d78141fc9648b3ef5edf7d

# Conflicts:
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/GsonTypeAdapterUtil.java
#	src/main/java/com/google/devtools/build/lib/bazel/bzlmod/SingleExtensionEvalFunction.java

* fixes

* Fix tests

* Enable lockfile by default and fix warning

PiperOrigin-RevId: 545927412
Change-Id: I34e8531b8e396ccdfe0eecbee22cb68f76f969fc

# Conflicts:
#	src/test/java/com/google/devtools/build/lib/analysis/RunfilesRepoMappingManifestTest.java
#	src/test/java/com/google/devtools/build/lib/bazel/bzlmod/BazelDepGraphFunctionTest.java
#	src/test/java/com/google/devtools/build/lib/bazel/bzlmod/BazelModuleResolutionFunctionTest.java
#	src/test/java/com/google/devtools/build/lib/query2/testutil/SkyframeQueryHelper.java
#	src/test/java/com/google/devtools/build/lib/rules/starlarkdocextract/StarlarkDocExtractTest.java

* Add lockfile documentation

PiperOrigin-RevId: 546085523
Change-Id: If287e6e143f8858d185f83a1630275520db65e44

# Conflicts:
#	site/en/_book.yaml

* Fix null event issue

If a change "only" occurred in a module extension, then the value of ModuleResolutionEvent is null -the skyvalue is cached and the event was never sent- then "combineModuleExtensions" function would crash with null pointer exception while trying to get the old usages.

In this case we can just re-add the old module extensions from the lockfile because if the module resolution is the same, this means the usage didn't change.

PiperOrigin-RevId: 546821911
Change-Id: Ie685cbab654d1c41403aebd31ddad91033be1d56

* Fix tests

* Remove unrelated updates

* reverse unrelated changes