Avoid overcounting in ActionLookupValueProgressReceiver.
Skips counting ConfiguredTargets or Actions for nodes that perform delegation.

Shifts proxy ConfiguredTarget node filtering from AnalysisTestCase to
BuildViewForTesting so the consistency check in
BuildViewForTesting.getSkyframeEvaluatedActionLookupKeyCountForTesting
against SkyFrameBuildView.getEvaluatedCounts passes. The check here is more
involved because it needs to handle nodes that are not done.

Not-done nodes occur when a subsequent incremental evaluation dirties, but
doesn't request a node.

PiperOrigin-RevId: 547192963
Change-Id: Ica01f20440ab0ba258e73c3f9cec5c7f92c99fa9