starlark_doc_extract: document repository rules and module extensions
We introduce new message types in the proto for repository rules and for
module extensions. Note that this is a behavioral change compared to the
legacy Stardoc extractor, which extracted repository rule docs using RuleInfo
messages (i.e. same as ordinary rules) - which was very likely to lead to
reader confusion.

ModuleExtensionInfo messages for now get an OriginKey field without the original
name because ModuleExtension objects are not StarlarkExportable values, so
finding the name under which a ModuleExtension was originally defined means
searching through its defining module's globals (assuming it is in globals - it
may not be). This will be added as a follow-up, but is low priority (due to the
lower probability of name collisions for module extensions as compared to e.g.
providers and rules) and should not block the ability to document module
extensions in the first place.

PiperOrigin-RevId: 547496231
Change-Id: I1ecb7f9b8bf2b7be78e565cb03a57d7227cda58f