Report dev/non-dev deps imported via non-dev/dev usages
If a repository generated by a module extension corresponds to a dev tag but is imported on a non-dev usage proxy (or vice versa), the build can fail if the module is used by other modules. `ModuleExtensionMetadata` already generated fixup commands for this case, but now also explains the situation and impact.

Closes #18832.

PiperOrigin-RevId: 547589845
Change-Id: I58a63b1aa6b60849a350202097515c7840540800