Raise an error when rule doesn't define java toolchain but uses java_common
Some functions inside java_common don't have `ctx` as a parameter. They do have `actions` parameter, from which is possible to get ctx in Native, but not in Starlark. In order not to add ctx to these functions, I've exposed native function which raises this error. Still, this check will be deleted once AEGs are fully enabled since then rules with get error directly from AEGs and will not add more Native code which could not be starlarkified.

PiperOrigin-RevId: 547600842
Change-Id: I1575e614fef092c7245baaaa80ae4e3302851237