Add the experimental_remote_require_cached flag
When set to true, this flag causes Bazel to abort the build whenever it
encounters an action that is not cached.

This is very useful when trying to troubleshoot action caching issues
across machines because it allows running a build on one and having it
fail on another as soon as there is a problem without tainting what
already exists in the cache.

My workflow is to essentially do:

1. Machine 1: bazel clean
2. Machine 1: bazel build ...
3. Machine 2: bazel clean
4. Machine 2: bazel build --experimental_remote_require_cached ...

which makes step 4 fail on the first action that wasn't cached as
expected. Then I can address that problem and re-run step 4 to
encounter the next issue.